# Base64 acceleration for ARM64