--- a/base64/base64_arm64.go
+++ b/base64/base64_arm64.go
@@ -3,11 +3,13 @@
 
 package base64
 
-import "encoding/base64"
+import (
+	"encoding/base64"
+)
 
 const (
 	encLutSize   = 16
-	decLutSize   = 48
+	decLutSize   = 2
 	minEncodeLen = 16 * 3
 	minDecodeLen = 8 * 4
 )
@@ -15,6 +17,7 @@ const (
 func newEncoding(encoder string) *Encoding {
 	e := &Encoding{base: base64.NewEncoding(encoder)}
 	e.enableEncodeARM64(encoder)
+	e.enableDecodeARM64(encoder)
 	return e
 }
 
@@ -27,4 +30,13 @@ func (e *Encoding) enableEncodeARM64(encoder string) {
 
 	e.enc = encodeARM64
 	e.enclut = tab
+}
+
+func (e *Encoding) enableDecodeARM64(encoder string) {
+	if encoder == encodeStd {
+		e.dec = decodeStdARM64
+	} else {
+		e.dec = decodeARM64
+	}
+	e.declut = [decLutSize]int8{int8(encoder[62]), int8(encoder[63])}
 }