--- a/base64/base64_default.go
+++ b/base64/base64_default.go
@@ -1,5 +1,5 @@
-//go:build purego || !amd64
-// +build purego !amd64
+//go:build purego || !(amd64 || arm64)
+// +build purego !amd64,!arm64
 
 package base64