--- a/base64/encode_arm64.go
+++ b/base64/encode_arm64.go
@@ -1,0 +1,6 @@
+//go:build !purego
+// +build !purego
+
+package base64
+
+func encodeARM64(dst []byte, src []byte, lut *int8) (int, int)