This PR introduces ARM64 support for base64.

The general purpose portion of `base64_amd64.go` was moved into `base64_asm.go` while keeping the `AMD64`-specific code isolated. A new `base64_arm64.go` implementation in being introduced to the current implementation.

On a Graviton 2 instance, the benchmarks are looking decent:
```goos: linux
goarch: arm64
pkg: github.com/segmentio/asm/base64
BenchmarkEncode/asm-4            1207927               994.1 ns/op      4120.41 MB/s
BenchmarkEncode/go-4              208002              5759 ns/op         711.18 MB/s
BenchmarkDecode/asm-4             534616              2242 ns/op        2437.53 MB/s
BenchmarkDecode/go-4              233426              5140 ns/op        1063.06 MB/s
```